refactor: extract modal components from exercise library page
