+++ app/(admin)/programs/[id]/exercises/page.tsx
@@ -9,12 +9,12 @@ import { Label } from '@/components/ui/label';
 import { Select } from '@/components/ui/select';
 import { Table } from '@/components/ui/table';
 import { Badge } from '@/components/ui/badge';
-import { Dialog } from '@/components/ui/dialog';
-import { Textarea } from '@/components/ui/textarea';
 import { Plus, Search, Edit, Trash2, Eye, Settings } from 'lucide-react';
 import { toast } from 'sonner';
 import { apiFetchJson } from '@/lib/api-helpers';
 import type { Exercise, TemplateExercise } from '@/lib/types';
+import ExerciseFormModal from './ExerciseFormModal';
+import WeightEditorModal from './WeightEditorModal';
 
 const MUSCLE_GROUPS = [
   'Chest',
@@ -351,117 +351,19 @@ export default function ExerciseLibraryPage() {
 
       {/* Create/Edit Exercise Modal */}
       {(showCreateModal || showEditModal) && (
-        <Dialog open={true} onOpenChange={() => {
-          setShowCreateModal(false);
-          setShowEditModal(false);
-          setSelectedExercise(null);
-          resetForm();
-        }}>
-          <div className="fixed inset-0 bg-black/50 flex items-center justify-center p-4">
-            <div className="bg-elevated rounded-lg border border-subtle p-6 w-full max-w-lg">
-              <h2 className="text-lg font-bold text-content-primary mb-4">
-                {showCreateModal ? 'Add New Exercise' : 'Edit Exercise'}
-              </h2>
-              
-              <div className="space-y-4">
-                <div>
-                  <Label htmlFor="name">Name *</Label>
-                  <Input
-                    id="name"
-                    value={exerciseForm.name}
-                    onChange={(e) => setExerciseForm(prev => ({ ...prev, name: e.target.value }))}
-                    placeholder="Exercise name"
-                  />
-                </div>
-
-                <div>
-                  <Label htmlFor="muscle-group-form">Muscle Group</Label>
-                  <Select
-                    value={exerciseForm.muscle_group}
-                    onChange={(e) => setExerciseForm(prev => ({ ...prev, muscle_group: e.target.value }))}
-                  >
-                    <option value="">Select muscle group</option>
-                    {MUSCLE_GROUPS.map(group => (
-                      <option key={group} value={group}>{group}</option>
-                    ))}
-                  </Select>
-                </div>
-
-                <div>
-                  <Label htmlFor="equipment">Equipment</Label>
-                  <Input
-                    id="equipment"
-                    value={exerciseForm.equipment}
-                    onChange={(e) => setExerciseForm(prev => ({ ...prev, equipment: e.target.value }))}
-                    placeholder="e.g., Barbell, Dumbbells, Bodyweight"
-                  />
-                </div>
-
-                <div>
-                  <Label htmlFor="instructions">Instructions</Label>
-                  <Textarea
-                    id="instructions"
-                    value={exerciseForm.instructions}
-                    onChange={(e) => setExerciseForm(prev => ({ ...prev, instructions: e.target.value }))}
-                    placeholder="Step-by-step exercise instructions"
-                    rows={3}
-                  />
-                </div>
-
-                <div>
-                  <Label htmlFor="coaching-cues">Coaching Cues</Label>
-                  <Textarea
-                    id="coaching-cues"
-                    value={exerciseForm.coaching_cues}
-                    onChange={(e) => setExerciseForm(prev => ({ ...prev, coaching_cues: e.target.value }))}
-                    placeholder="Key coaching points and form cues"
-                    rows={2}
-                  />
-                </div>
-
-                <div>
-                  <Label htmlFor="video-url">Video URL</Label>
-                  <Input
-                    id="video-url"
-                    type="url"
-                    value={exerciseForm.video_url}
-                    onChange={(e) => setExerciseForm(prev => ({ ...prev, video_url: e.target.value }))}
-                    placeholder="https://..."
-                  />
-                </div>
-
-                {showEditModal && (
-                  <div className="flex items-center gap-2">
-                    <input
-                      type="checkbox"
-                      id="is-active"
-                      checked={exerciseForm.is_active}
-                      onChange={(e) => setExerciseForm(prev => ({ ...prev, is_active: e.target.checked }))}
-                    />
-                    <Label htmlFor="is-active">Active</Label>
-                  </div>
-                )}
-              </div>
-
-              <div className="flex justify-end gap-3 mt-6">
-                <Button 
-                  variant="ghost" 
-                  onClick={() => {
-                    setShowCreateModal(false);
-                    setShowEditModal(false);
-                    setSelectedExercise(null);
-                    resetForm();
-                  }}
-                >
-                  Cancel
-                </Button>
-                <Button onClick={showCreateModal ? handleCreateExercise : handleUpdateExercise}>
-                  {showCreateModal ? 'Add Exercise' : 'Save Changes'}
-                </Button>
-              </div>
-            </div>
-          </div>
-        </Dialog>
+        <ExerciseFormModal
+          mode={showCreateModal ? 'create' : 'edit'}
+          exerciseForm={exerciseForm}
+          onFormChange={setExerciseForm}
+          onSubmit={showCreateModal ? handleCreateExercise : handleUpdateExercise}
+          onClose={() => {
+            setShowCreateModal(false);
+            setShowEditModal(false);
+            setSelectedExercise(null);
+            resetForm();
+          }}
+          MUSCLE_GROUPS={MUSCLE_GROUPS}
+        />
       )}
 
       {/* Weight Editor Modal - We'll implement this next */}
@@ -483,232 +385,3 @@ export default function ExerciseLibraryPage() {
   );
 }
 
-// Weight Editor Modal Component
-interface WeightEditorModalProps {
-  templateExercise: TemplateExercise;
-  onClose: () => void;
-  onSave: () => void;
-}
-
-function WeightEditorModal({ templateExercise, onClose, onSave }: WeightEditorModalProps) {
-  const [formData, setFormData] = useState({
-    sets_default: templateExercise.sets_default,
-    reps_default: templateExercise.reps_default,
-    weight_pre_baseline_f: templateExercise.weight_pre_baseline_f,
-    weight_pre_baseline_m: templateExercise.weight_pre_baseline_m,
-    weight_default_f: templateExercise.weight_default_f,
-    weight_default_m: templateExercise.weight_default_m,
-    weight_post_baseline_f: templateExercise.weight_post_baseline_f,
-    weight_post_baseline_m: templateExercise.weight_post_baseline_m,
-    is_bodyweight: templateExercise.is_bodyweight,
-    superset_group: templateExercise.superset_group || '',
-    coaching_cues: templateExercise.coaching_cues || '',
-    notes: templateExercise.notes || ''
-  });
-
-  const handleSave = async () => {
-    try {
-      await apiFetchJson('/api/admin/exercises/template-exercises', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ id: templateExercise.id, ...formData })
-      });
-      onSave();
-      toast.success('Weight settings saved');
-    } catch {
-      toast.error('Failed to save weight settings');
-    }
-  };
-
-  return (
-    <Dialog open={true} onOpenChange={onClose}>
-      <div className="fixed inset-0 bg-black/50 flex items-center justify-center p-4">
-        <div className="bg-elevated rounded-lg border border-subtle p-6 w-full max-w-4xl max-h-[90vh] overflow-y-auto">
-          <h2 className="text-lg font-bold text-content-primary mb-4">
-            Weight Editor: {templateExercise.exercise?.name}
-          </h2>
-          
-          <div className="grid gap-6">
-            {/* Sets and Reps */}
-            <div className="grid grid-cols-2 gap-4">
-              <div>
-                <Label htmlFor="sets">Default Sets</Label>
-                <Input
-                  id="sets"
-                  type="number"
-                  min="1"
-                  value={formData.sets_default}
-                  onChange={(e) => setFormData(prev => ({ ...prev, sets_default: parseInt(e.target.value) || 1 }))}
-                />
-              </div>
-              <div>
-                <Label htmlFor="reps">Default Reps</Label>
-                <Input
-                  id="reps"
-                  type="number"
-                  min="1"
-                  value={formData.reps_default}
-                  onChange={(e) => setFormData(prev => ({ ...prev, reps_default: parseInt(e.target.value) || 1 }))}
-                />
-              </div>
-            </div>
-
-            {/* Weight Defaults Grid */}
-            <div>
-              <h3 className="font-medium text-content-primary mb-3">Weight Defaults (lbs)</h3>
-              <div className="grid grid-cols-3 gap-4">
-                <div className="text-center">
-                  <div className="font-medium text-content-primary mb-2">Pre-Baseline</div>
-                  <div className="space-y-2">
-                    <div>
-                      <Label htmlFor="weight-pre-f" className="text-sm">Female</Label>
-                      <Input
-                        id="weight-pre-f"
-                        type="number"
-                        min="0"
-                        step="2.5"
-                        value={formData.weight_pre_baseline_f}
-                        onChange={(e) => setFormData(prev => ({ ...prev, weight_pre_baseline_f: parseFloat(e.target.value) || 0 }))}
-                        disabled={formData.is_bodyweight}
-                      />
-                    </div>
-                    <div>
-                      <Label htmlFor="weight-pre-m" className="text-sm">Male</Label>
-                      <Input
-                        id="weight-pre-m"
-                        type="number"
-                        min="0"
-                        step="2.5"
-                        value={formData.weight_pre_baseline_m}
-                        onChange={(e) => setFormData(prev => ({ ...prev, weight_pre_baseline_m: parseFloat(e.target.value) || 0 }))}
-                        disabled={formData.is_bodyweight}
-                      />
-                    </div>
-                  </div>
-                </div>
-
-                <div className="text-center">
-                  <div className="font-medium text-content-primary mb-2">Default</div>
-                  <div className="space-y-2">
-                    <div>
-                      <Label htmlFor="weight-default-f" className="text-sm">Female</Label>
-                      <Input
-                        id="weight-default-f"
-                        type="number"
-                        min="0"
-                        step="2.5"
-                        value={formData.weight_default_f}
-                        onChange={(e) => setFormData(prev => ({ ...prev, weight_default_f: parseFloat(e.target.value) || 0 }))}
-                        disabled={formData.is_bodyweight}
-                      />
-                    </div>
-                    <div>
-                      <Label htmlFor="weight-default-m" className="text-sm">Male</Label>
-                      <Input
-                        id="weight-default-m"
-                        type="number"
-                        min="0"
-                        step="2.5"
-                        value={formData.weight_default_m}
-                        onChange={(e) => setFormData(prev => ({ ...prev, weight_default_m: parseFloat(e.target.value) || 0 }))}
-                        disabled={formData.is_bodyweight}
-                      />
-                    </div>
-                  </div>
-                </div>
-
-                <div className="text-center">
-                  <div className="font-medium text-content-primary mb-2">Post-Baseline</div>
-                  <div className="space-y-2">
-                    <div>
-                      <Label htmlFor="weight-post-f" className="text-sm">Female</Label>
-                      <Input
-                        id="weight-post-f"
-                        type="number"
-                        min="0"
-                        step="2.5"
-                        value={formData.weight_post_baseline_f}
-                        onChange={(e) => setFormData(prev => ({ ...prev, weight_post_baseline_f: parseFloat(e.target.value) || 0 }))}
-                        disabled={formData.is_bodyweight}
-                      />
-                    </div>
-                    <div>
-                      <Label htmlFor="weight-post-m" className="text-sm">Male</Label>
-                      <Input
-                        id="weight-post-m"
-                        type="number"
-                        min="0"
-                        step="2.5"
-                        value={formData.weight_post_baseline_m}
-                        onChange={(e) => setFormData(prev => ({ ...prev, weight_post_baseline_m: parseFloat(e.target.value) || 0 }))}
-                        disabled={formData.is_bodyweight}
-                      />
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            {/* Additional Settings */}
-            <div className="grid grid-cols-2 gap-4">
-              <div>
-                <div className="flex items-center gap-2 mb-3">
-                  <input
-                    type="checkbox"
-                    id="is-bodyweight"
-                    checked={formData.is_bodyweight}
-                    onChange={(e) => setFormData(prev => ({ ...prev, is_bodyweight: e.target.checked }))}
-                  />
-                  <Label htmlFor="is-bodyweight">Bodyweight Exercise</Label>
-                </div>
-                
-                <div>
-                  <Label htmlFor="superset">Superset Group</Label>
-                  <Input
-                    id="superset"
-                    value={formData.superset_group}
-                    onChange={(e) => setFormData(prev => ({ ...prev, superset_group: e.target.value }))}
-                    placeholder="e.g., A1, B2"
-                  />
-                </div>
-              </div>
-
-              <div className="space-y-3">
-                <div>
-                  <Label htmlFor="exercise-cues">Exercise-Specific Cues</Label>
-                  <Textarea
-                    id="exercise-cues"
-                    value={formData.coaching_cues}
-                    onChange={(e) => setFormData(prev => ({ ...prev, coaching_cues: e.target.value }))}
-                    placeholder="Specific cues for this template"
-                    rows={2}
-                  />
-                </div>
-                
-                <div>
-                  <Label htmlFor="exercise-notes">Notes</Label>
-                  <Textarea
-                    id="exercise-notes"
-                    value={formData.notes}
-                    onChange={(e) => setFormData(prev => ({ ...prev, notes: e.target.value }))}
-                    placeholder="Internal notes"
-                    rows={2}
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-
-          <div className="flex justify-end gap-3 mt-6">
-            <Button variant="ghost" onClick={onClose}>
-              Cancel
-            </Button>
-            <Button onClick={handleSave}>
-              Save Weight Settings
-            </Button>
-          </div>
-        </div>
-      </div>
-    </Dialog>
-  );
-}
